Refactor and modularize app providers into separate files for Modeler, Autosave, and Import features. Streamline `app.config.ts` by replacing inline initializers with provider functions.
