+++ src/app/tools/autosave/autosave.providers.ts
@@ -1,0 +1,8 @@
+import { inject, provideAppInitializer } from '@angular/core';
+import { AutosaveService }               from './services/autosave.service';
+
+export function provideAutosave() {
+  return provideAppInitializer(() => {
+    inject(AutosaveService);
+  });
+}
